Refactor process management and enhance shell commands
- Updated process handling in `process.c` to ensure proper type consistency for process IDs.
- Modified `process.h` to include standard types and adjusted the `fd_entry_t` structure for offset handling.
- Cleaned up `task.h` by removing unnecessary comments and organizing fields for better readability.
- Enhanced `shell.c` with new commands: `whoami`, `cp`, `mv`, `chmod`, and `df`, providing improved file management capabilities.
- Implemented robust output redirection handling in `cmd_echo` and ensured safe file operations.
- Added `lsblk` command to list block devices and their partitions, improving disk management visibility.
- Introduced user management functionality with a simple root user implementation in `user.c` and `user.h`.
- Updated syscall handling in `syscalls.c` to support new commands and improved error reporting.
- Refined syscall definitions in `syscalls.h` and `syscalls_internal.h` for better type consistency.
- Adjusted user-related types in `types.h` and `unistd.h` for compatibility with the new user subsystem.

+++ kernel/process.c
@@ -240,8 +240,8 @@ int process_wait(pid_t pid, int *status)
 		p = process_list;
 		while (p)
 		{
-			if (p->task.pid == pid &&
-			    p->ppid == current_process->task.pid)
+			// Both pids are now of type pid_t (int32_t)
+			if ((uint32_t)p->task.pid == (uint32_t)pid && (uint32_t)p->ppid == (uint32_t)current_process->task.pid) // Conversión explícita
 			{
 				if (p->state == PROCESS_ZOMBIE)
 				{
